# tests/openapi.yaml
openapi: 3.0.3
info:
  title: Tiikii Festival API
  description: API para el sistema de festivales Tiikii con funcionalidades de tickets, chat, mesh network y más
  version: 1.0.0
  contact:
    name: Tiikii Development Team
    email: dev@tiikii.com
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
  - url: https://api.tiikii.com
    description: Servidor de producción
security:
  - BearerAuth: []
paths:
...
  /api/artists:
    get:
      summary: Obtener todos los artistas
      description: Lista todos los artistas con filtros opcionales
      parameters:
        - name: genre
          in: query
          schema:
            type: string
          example: "electronic"
        - name: festivalId
          in: query
          schema:
            type: string
          example: "festival-123"
      responses:
        '200':
          description: Lista de artistas
          content:
            application/json:
              schema:
                type: object
                properties:
                  artists:
                    type: array
                    items:
                      $ref: '#/components/schemas/Artist'
                  count:
                    type: integer
  /api/artists/{id}:
    get:
      summary: Obtener artista por ID
      description: Obtiene información detallada de un artista específico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "artist-123"
      responses:
        '200':
          description: Información del artista
          content:
            application/json:
              schema:
                type: object
                properties:
                  artist:
                    $ref: '#/components/schemas/ArtistDetail'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/artists/genre/{genre}:
    get:
      summary: Obtener artistas por género
      description: Lista artistas filtrados por género
      parameters:
        - name: genre
          in: path
          required: true
          schema:
            type: string
          example: "electronic"
        - name: festivalId
          in: query
          schema:
            type: string
          example: "festival-123"
      responses:
        '200':
          description: Lista de artistas por género
          content:
            application/json:
              schema:
                type: object
                properties:
                  genre:
                    type: string
                  artists:
                    type: array
                    items:
                      $ref: '#/components/schemas/Artist'
                  count:
                    type: integer
  /api/artists/search/{query}:
    get:
      summary: Buscar artistas
      description: Busca artistas por nombre, biografía o género
      parameters:
        - name: query
          in: path
          required: true
          schema:
            type: string
          example: "deadmau5"
        - name: festivalId
          in: query
          schema:
            type: string
          example: "festival-123"
      responses:
        '200':
          description: Resultados de búsqueda
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  festivalId:
                    type: string
                  artists:
                    type: array
                    items:
                      $ref: '#/components/schemas/Artist'
                  count:
                    type: integer
  /api/artists/festival/{festivalId}:
    get:
      summary: Obtener artistas de festival
      description: Lista todos los artistas de un festival específico
      parameters:
        - name: festivalId
          in: path
          required: true
          schema:
            type: string
          example: "festival-123"
        - name: genre
          in: query
          schema:
            type: string
          example: "electronic"
      responses:
        '200':
          description: Lista de artistas del festival
          content:
            application/json:
              schema:
                type: object
                properties:
                  festivalId:
                    type: string
                  festivalName:
                    type: string
                  artists:
                    type: array
                    items:
                      $ref: '#/components/schemas/Artist'
                  count:
                    type: integer
  /api/artists/stats/overview:
    get:
      summary: Estadísticas de artistas
      description: Obtiene estadísticas generales de artistas
      parameters:
        - name: festivalId
          in: query
          schema:
            type: string
          example: "festival-123"
      responses:
        '200':
          description: Estadísticas de artistas
          content:
            application/json:
              schema:
                type: object
                properties:
                  festivalId:
                    type: string
                  stats:
                    type: object
                    properties:
                      totalArtists:
                        type: integer
                      totalGenres:
                        type: integer
                  genres:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        count:
                          type: integer
                  artistsByGenre:
                    type: array
                    items:
                      type: object
                      properties:
                        genre:
                          type: string
                        artistCount:
                          type: integer
                        performanceCount:
                          type: integer
  /api/artists/popular:
    get:
      summary: Obtener artistas populares
      description: Lista los artistas más populares por número de presentaciones
      parameters:
        - name: festivalId
          in: query
          schema:
            type: string
          example: "festival-123"
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          example: 5
      responses:
        '200':
          description: Lista de artistas populares
          content:
            application/json:
              schema:
                type: object
                properties:
                  festivalId:
                    type: string
                  limit:
                    type: integer
                  artists:
                    type: array
                    items:
                      $ref: '#/components/schemas/PopularArtist'
  /api/artists/{id}/upcoming:
    get:
      summary: Obtener presentaciones próximas del artista
      description: Lista las próximas presentaciones de un artista
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "artist-123"
      responses:
        '200':
          description: Próximas presentaciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  artist:
                    $ref: '#/components/schemas/Artist'
                  upcomingPerformances:
                    type: array
                    items:
                      $ref: '#/components/schemas/Performance'
                  count:
                    type: integer
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Artist:
      type: object
      properties:
        id:
          type: string
          example: "artist-123"
        name:
          type: string
          example: "Deadmau5"
        bio:
          type: string
          example: "Canadian electronic music producer and DJ"
        genre:
          type: string
          example: "electronic"
        imageUrl:
          type: string
          example: "https://example.com/artist.jpg"
        socialMedia:
          type: object
          example: {"twitter": "@deadmau5", "instagram": "@deadmau5"}
    ArtistDetail:
      allOf:
        - $ref: '#/components/schemas/Artist'
        - type: object
          properties:
            schedule:
              type: array
              items:
                $ref: '#/components/schemas/Performance'
    PopularArtist:
      allOf:
        - $ref: '#/components/schemas/Artist'
        - type: object
          properties:
            performanceCount:
              type: integer
              example: 3
    Performance:
      type: object
      properties:
        id:
          type: string
          example: "performance-123"
        artistId:
          type: string
          example: "artist-123"
        festivalId:
          type: string
          example: "festival-123"
        stageId:
          type: string
          example: "stage-123"
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        festival:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            logo:
              type: string
        stage:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            description:
              type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            type:
              type: string
              example: "ValidationError"
            message:
              type: string
              example: "Invalid input data"
            code:
              type: string
              example: "VALIDATION_ERROR"
            field:
              type: string
              example: "email"
            details:
              type: array
              items:
                type: string
              example: ["Field 'email' is required"]
        timestamp:
          type: string
          format: date-time
  responses:
...
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: "NotFoundError"
              message: "Resource not found"
              code: "NOT_FOUND"
...